--- app/oauth2/success/page.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import {requestGet} from '@/shared/apis';
-import {useEffect} from 'react';
-
-export default function Page() {
-  useEffect(() => {
-    const tmp = async () => {
-      await requestGet({
-        endpoint: '/hello',
-      });
-    };
-
-    tmp();
-  }, []);
-  return <section>tmp</section>;
-}
