+++ src/features/oauth2/index.ts
@@ -1,0 +1,1 @@
+export {default as useKakaoOAuthLogin} from './lib/useKakaoOAuthLogin';
