+++ src/views/oauth2/ui/OAuth2RedirectPage.tsx
@@ -1,36 +1,10 @@
 'use client';
 
-import {useEffect} from 'react';
-import {useRouter, useSearchParams} from 'next/navigation';
-import {requestGet} from '@/shared/apis';
-import {createClientError} from '@/shared/lib/utils/client-error';
+import {useKakaoOAuthLogin} from '@/features/oauth2';
 import {LucideIcon} from '@/shared/ui/icons';
 
 export default function OAuth2RedirectPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-
-  useEffect(() => {
-    const email = searchParams.get('user_email');
-
-    if (!email) {
-      throw createClientError('EMPTY_USER_EMAIL');
-    }
-
-    const kakaoLogin = async () => {
-      requestGet({
-        endpoint: '/user/oauth2/login',
-        queryParams: {
-          user_email: email,
-        },
-        withResponse: false,
-      });
-
-      router.push('/');
-    };
-
-    kakaoLogin();
-  }, [router, searchParams]);
+  useKakaoOAuthLogin();
 
   return (
     <section className="w-full h-full flex flex-col justify-center items-center">
